requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
konlpy==0.6.0
scikit-learn==1.5.0
numpy==1.26.4
pymysql==1.1.1
python-dotenv==1.0.1 
